% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesQuery.R
\name{caracsite}
\alias{caracsite}
\title{caracsite}
\usage{
caracsite(pool, language)
}
\arguments{
\item{pool}{data base configuration (from confConnexion function)}

\item{language}{String of language ("fr" or "en")}
}
\value{
data.table characteristics of sites
}
\description{
Sql query to extrat characteristics of sites
}
