% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{checkboxVariablePedo}
\alias{checkboxVariablePedo}
\title{checkboxVariablePedo}
\usage{
checkboxVariablePedo(dbCarac, codeVariable, siteBM)
}
\arguments{
\item{dbCarac}{data.table from mod_accessdata}

\item{codeVariable}{String of soil variables (ex. "TS_1_1_1")}

\item{siteBM}{: String of biometeorological site/station code selected in application. Ex. "lgt/bm1" ou c("lgt/bm2","lgt/bm1")}
}
\value{
data.table with a simple definition (without soil profil definition)
}
\description{
Create a list of variable with a simple definition (without soil profil definition)
}
