% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesQuery.R
\name{confConnexion}
\alias{confConnexion}
\title{confConnexion}
\usage{
confConnexion(pathConfFile)
}
\arguments{
\item{pathConfFile}{path of yaml file database configuration}
}
\value{
pool configuration
}
\description{
Connexion configuration to database
}
\examples{
\dontrun{
pool <- confConnexion("~/SI_SNOT/dbconfProd.yaml")
}
}
