% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{createFilePivot}
\alias{createFilePivot}
\title{createFilePivot}
\usage{
createFilePivot(data, siteVariable, caracDataset)
}
\arguments{
\item{data}{data to be integrated into a file with Pivot/TheiaOzcar format}

\item{siteVariable}{data.table with code_site_station and a variable}

\item{caracDataset}{table of dataset characteristics (create with caracdata function)}
}
\value{
Retourne une data.table au format Pivot/TheiaOzcar
}
\description{
Function to create data in Pivot/TheiaOzcar format
}
