% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{createLink}
\alias{createLink}
\title{createLink}
\usage{
createLink(val)
}
\arguments{
\item{val}{String of url (ex. "https://data-snot.cnrs.fr/")}
}
\value{
String of HTML code of url with '_blank' target configuration
}
\description{
To creat an html link
}
\examples{
createLink("https://data-snot.cnrs.fr/")
}
