% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{dbDayNight}
\alias{dbDayNight}
\title{dbDayNight}
\usage{
dbDayNight(dataSelected)
}
\arguments{
\item{dataSelected}{data.table from mod_accessdata}
}
\value{
day and night parameter into 'dataSelected' data.table
}
\description{
Create day and night parameter into 'dataSelected' data.table
}
