% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{dyBarChart}
\alias{dyBarChart}
\title{dyBarChart}
\source{
https://rstudio.github.io/dygraphs/gallery-custom-plotters.html
}
\usage{
dyBarChart(dygraph)
}
\arguments{
\item{dygraph}{dygraph object}
}
\value{
css configuration for dygraph
}
\description{
Bar chart in dygraph
}
\examples{
\dontrun{
dygraph(data)\%>\%
dyBarChart()
}
}
