% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{dyCSScool}
\alias{dyCSScool}
\title{dyCSScool}
\source{
https://github.com/rstudio/dygraphs/issues/227
}
\usage{
dyCSScool(dygraph)
}
\arguments{
\item{dygraph}{dygraph object}
}
\value{
css configuration for dygraph
}
\description{
css configuration for dygraph
}
\examples{
\dontrun{
dygraph(data)\%>\%
dyCSScool()
}
}
