% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesGraph.R
\name{dygraphTypeVariable}
\alias{dygraphTypeVariable}
\title{dygraphTypeVariable}
\usage{
dygraphTypeVariable(subsetoutbd, frequence)
}
\arguments{
\item{subsetoutbd}{data.table from mod_accessdata}

\item{frequence}{: String of frequency (ex. "30 min", "hour", "day", "month")}
}
\value{
dygraph of timetrend
}
\description{
Time trend with dygraphs library for different type of variable
}
