% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{dyMultiColumn}
\alias{dyMultiColumn}
\title{dyMultiColumn}
\source{
https://rstudio.github.io/dygraphs/gallery-custom-plotters.html
}
\usage{
dyMultiColumn(dygraph)
}
\arguments{
\item{dygraph}{dygraph object}
}
\value{
css configuration for dygraph
}
\description{
MultiColumn in dygraph
}
\examples{
\dontrun{
dygraph(data)\%>\%
dyMultiColumn()
}
}
