% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{extracData}
\alias{extracData}
\title{extracData}
\usage{
extracData(db, frequence)
}
\arguments{
\item{db}{data.table from mod_accessdata}

\item{frequence}{String of frequency (ex. "30 min", "hour", "day", "month")}
}
\value{
data.table with horizontal format
}
\description{
Fonction used to export data with horizontal format (for download it)
}
