% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{googleAnalyticsParameter}
\alias{googleAnalyticsParameter}
\title{googleAnalyticsParameter}
\usage{
googleAnalyticsParameter()
}
\value{
String of java script
}
\description{
googleAnalytics paramters
}
\examples{
\dontrun{
googleAnalyticsParameter()
}
}
