% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{grid_arrange_shared_legend}
\source{
http://stackoverflow.com/questions/13649473/add-a-common-legend-for-combined-ggplots
}
\usage{
grid_arrange_shared_legend(
  ...,
  nrow = 1,
  ncol = length(list(...)),
  position = c("bottom", "right")
)
}
\arguments{
\item{...}{ggplot objects can be passed in ...}

\item{nrow}{ggplot parameter}

\item{ncol}{ggplot parameter}

\item{position}{legend position ("bottom", "right","top","left")}
}
\value{
ggplots
}
\description{
Combined ggplots with a common legend
}
