% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{loadCaracDataAndCarto}
\alias{loadCaracDataAndCarto}
\title{loadCaracDataAndCarto}
\usage{
loadCaracDataAndCarto(pool, language)
}
\arguments{
\item{pool}{data base configuration (from confConnexion function)}

\item{language}{String of language ("fr" or "en")}
}
\value{
3 data.table
}
\description{
Load data caracdata() and create 3 data.table used in application
}
