% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesQuery.R
\name{queryDataSNOT}
\alias{queryDataSNOT}
\title{queryDataSNOT}
\usage{
queryDataSNOT(
  pool,
  variableSelected,
  siteSelected,
  periodeSelected,
  melted = TRUE
)
}
\arguments{
\item{pool}{data base configuration (from confConnexion function)}

\item{variableSelected}{String of variable code selected in application (ex. "FC" or c("FC","FCH4"))#'}

\item{siteSelected}{String of site/station code selected in application (ex. "lgt/ec1" or c("lgt/ec1","lgt/bm1"))}

\item{periodeSelected}{Vector string of period selected with "yyyy-mm-dd" format (ex. c("2008-01-15","2014-05-25"))}

\item{melted}{Logical. If TRUE convert melt data.table. If FALSE, no melt (used for dataset-archive)}
}
\value{
data.table of values
}
\description{
Sql query to extrat values of variable from site/station and period selected
}
