% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{robustCurl}
\alias{robustCurl}
\title{robustCurl}
\usage{
robustCurl(x)
}
\arguments{
\item{x}{a httr::GET(url)}
}
\value{
Error or warning message if httr::GET() don't work
}
\description{
tryCatch
}
\examples{
\dontrun{
robustCurl(httr::GET(
"http://localhost:8081/rest/resources/pivot?codes_jeu=lgt-ges-eddycovariance"),
httr::timeout(60)))
}
}
