% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesGraph.R
\name{sensorSelectedMap}
\alias{sensorSelectedMap}
\title{sensorSelectedMap}
\usage{
sensorSelectedMap(mapSensorSelected, allSite = FALSE, translator)
}
\arguments{
\item{mapSensorSelected}{data.table from mod_welcome}

\item{allSite}{Logical operator to select allSite or not (TRUE, FALSE)}

\item{translator}{translator object for translation}
}
\value{
leaflet
}
\description{
Chart of disponibility data (welcome module)
}
