% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesQuery.R
\name{tableJeu}
\alias{tableJeu}
\title{tableJeu}
\usage{
tableJeu(pool)
}
\arguments{
\item{pool}{data base configuration (from confConnexion function)}
}
\value{
data.table of table jeu
}
\description{
Sql query to extrat table jeu from database
}
