% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesQuery.R
\name{tableSousJeu}
\alias{tableSousJeu}
\title{tableSousJeu}
\usage{
tableSousJeu(pool)
}
\arguments{
\item{pool}{data base configuration (from confConnexion function)}
}
\value{
data.table of table sousjeu
}
\description{
Sql query to extrat table sousjeu from database
}
