% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{theme_chart}
\alias{theme_chart}
\title{theme_chart2}
\usage{
theme_chart(position = "bottom", ...)
}
\arguments{
\item{position}{String of position of legend (ex. "bottom")}

\item{...}{ggplot objects can be passed in ...}
}
\value{
ggplot2 theme configuration
}
\description{
ggplot2 theme
}
