% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesGraph.R
\name{timelineDataAvailable}
\alias{timelineDataAvailable}
\title{timelineDataAvailable}
\usage{
timelineDataAvailable(tableCarac, allSite, facetWrapOption, translator)
}
\arguments{
\item{tableCarac}{data.table from mod_welcome}

\item{allSite}{Logical operator to select allSite or not (TRUE, FALSE)}

\item{facetWrapOption}{Logical operator to use facetWrap on site or site/station (TRUE, FALSE)}

\item{translator}{translator object for translation}
}
\value{
dygraph of timetrend
}
\description{
Chart of data disponibility (used in mod_welcome)
}
