% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{tweaks}
\alias{tweaks}
\title{tweaks}
\source{
https://stackoverflow.com/questions/36898492/aligning-checkbox-elements-in-many-coumns-in-a-shiny-app
}
\usage{
tweaks()
}
\value{
String of java script
}
\description{
Align checkboxs on multiple columns
}
\examples{
\dontrun{
tweaks()
}
}
