% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{tweaks2}
\alias{tweaks2}
\title{tweaks2}
\source{
https://stackoverflow.com/questions/42742191/align-checkboxgroupinput-vertically-and-horizontally
}
\usage{
tweaks2()
}
\value{
String of java script
}
\description{
Align checkboxs on multiple columns
}
\examples{
\dontrun{
tweaks2()
}
}
