% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{writeHeaderFile}
\alias{writeHeaderFile}
\title{writeHeaderFile}
\source{
https://stackoverflow.com/questions/12381117/add-header-to-file-created-by-write-csv
}
\usage{
writeHeaderFile(x, file, header)
}
\arguments{
\item{x}{data.frame or data.table}

\item{file}{String of file name}

\item{header}{String of header}
}
\value{
Write a file with header
}
\description{
Create a file with header
}
